+++ sys/arch/mvme68k/dev/vmes.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: vmes.c,v 1.7 2001/11/01 12:13:46 art Exp $ */
+/*	$OpenBSD: vmes.c,v 1.8 2001/11/05 19:45:34 art Exp $ */
 
 /*
  * Copyright (c) 1995 Theo de Raadt
@@ -165,7 +165,7 @@ vmesmmap(dev, off, prot)
 	struct vmessoftc *sc = (struct vmessoftc *) vmes_cd.cd_devs[unit];
 	void * pa;
 
-	pa = vmepmap(sc->sc_vme, (void *)off, NBPG, BUS_VMES);
+	pa = vmepmap(sc->sc_vme, (void *)(int)off, NBPG, BUS_VMES);
 	if (pa == NULL)
 		return (-1);
 	return (m68k_btop(pa));
